--- webview-ui/src/Layout/Cloud/Resources/styles.module.css
@@ -1,23 +1,0 @@
-.title {
-  font-size: 1.1rem;
-  font-weight: 600;
-  margin-bottom: 1rem;
-  margin-top: 0;
-  color: var(--vscode-foreground);
-}
-
-.intro {
-  font-size: 12px;
-  line-height: 1.5;
-  margin-bottom: 1.5rem;
-  color: var(--vscode-foreground);
-}
-
-.list {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  & > li {
-    margin-bottom: 0.5rem;
-  }
-}
